% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{aliases}
\alias{aliases}
\alias{len}
\alias{reversed}
\alias{sorted}
\title{Python-inspired aliases for R functions}
\usage{
len()

reversed(x)

sorted(x, decreasing = FALSE, ...)
}
\description{
These are some aliases for R function which are
more in line with Python equivalents.
}
\examples{
# python abbreviates to len, R is lengthier
len(1:5)

# for rev the trend is reversed
reversed(5:1)

# sorted
sorted(sample(1:5))
}
