% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{by_line}
\alias{by_line}
\title{Iterate over a connection by line}
\usage{
by_line(connection, FUN, STOP, safe = TRUE, max_lines = 0, ...)
}
\arguments{
\item{connection}{an open connection, such as a
text connection as created by \code{\link[base]{file}}}

\item{FUN}{a function to apply to each line
that can be read from the connection}

\item{STOP}{a function to generate a stop condition
for reading lines of the connection}

\item{safe}{auto-stop iteraction on reaching an
empty line in the connection (attempts to prevent
infinite loops)}

\item{...}{extra arguments passed to \code{FUN}}
}
\description{
Apply a function line-by-line to a connection,
such as an open file connection. This allows
simple processing without loading the entire
file to memory, something like the common
\code{for line in file: ...} approach in python.
}
\examples{
\dontrun{
file <- file("myfile.txt", "r")

# use with for automatic connection closing
with(file, {
  # read and print each line sequentially
  by_line(file, print)
})

# stop reading after finding the word 'stop'
connex <- file("logfile.txt", "r")
by_line(connex, cat, function(line) grepl(line, "stop"))
close(connex)
}

}
