% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyfuns.R
\name{\%within\%}
\alias{\%within\%}
\title{Check for character within string}
\usage{
char \%within\% string
}
\arguments{
\item{char}{a character or substring to look
for}

\item{string}{a string in which to match the
character or substring}
}
\description{
Simple wrapper that checks if a character
or substring has an exact match in a string.
Really just a wrapper around \code{grepl}.
}
\examples{

if ("w" \%within\% "hello, world!")
  print("Found it!")

"char" \%within\% "character string"

}
