% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pythonistr.R
\docType{package}
\name{pythonistr}
\alias{pythonistr}
\alias{pythonistr-package}
\title{pythonistr: Python idioms for R.}
\description{
Brings some useful Python-inspired functions
and aliases to R, making it slightly easier to move
back and forth between to two languages.
}
