% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyfuns.R
\name{s}
\alias{s}
\title{Separate a string into words}
\usage{
s(string)
separate(string)
}
\arguments{
\item{string}{a spaced character string to be
split into a word vector}
}
\description{
Split a space-delimited string into
a character vector of words, this saves
a bit of typing and quote-matching similar
to \code{'list of strings'.split()} in
python.
}
\examples{
df <- data.frame(a = s("chr1 chr2 chrMT"),
                 b = 1:3,
                 c = 5:7)

colnames(df) <- s("chr start end")
df

# separate is an alias for s
separate("split this string")

}
