% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\docType{data}
\name{aliases}
\alias{aliases}
\alias{len}
\alias{reversed}
\alias{sorted}
\alias{True}
\alias{False}
\title{Python-inspired aliases for R functions}
\format{An object of class \code{logical} of length 1.}
\usage{
len()

reversed(x)

sorted(x, decreasing = FALSE, ...)

True

False
}
\description{
These are some aliases for R function which are
more in line with Python equivalents.
}
\examples{
# python abbreviates to len, R is lengthier
base::length(1:5)
len(1:5)

# for rev the trend is reversed
base::rev(5:1)
reversed(5:1)

# returns a sorted copy
base::sort(sample(1:5))
sorted(sample(1:5))

# don't really use these...
identical(False, FALSE)
isTRUE(True)
}
\keyword{datasets}
