% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyfuns.R
\name{with.connection}
\alias{with.connection}
\title{Pseudo context-manager for file connections}
\usage{
\method{with}{connection}(connection, expr)
}
\arguments{
\item{connection}{a file connection created with \code{\link[base]{file}}}

\item{expr}{expression to evaluate}
}
\description{
A \code{with} method for file connection objects
which closes the connection after evaluation. This is a
system in Python.
}
\examples{
\dontrun{
connection <- file("file.csv", "r")

with(connection, {
  print(connection)
})

isOpen(connection)
}

}
\seealso{
\code{\link[base]{with}}
}
\keyword{connection}
